[loong64][maglev] Fix ambiguous call to li with char_code
Related to commit 906d4339a01c7eb588705c0ea684b8c0029620b6

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7960323
Auto-Submit: Liu Yu <liuyu@loongson.cn>
Commit-Queue: Zhao Jiazhong <zhaojiazhong-hf@loongson.cn>
Cr-Commit-Position: refs/heads/main@{#108069}

+++ src/maglev/loong64/maglev-ir-loong64.cc
@@ -199,7 +199,7 @@ void BuiltinStringFromCharCode::GenerateCode(MaglevAssembler* masm,
       __ LoadSingleCharacterString(result_string, char_code);
     } else {
       __ AllocateTwoByteString(register_snapshot(), result_string, 1);
-      __ li(scratch, char_code);
+      __ li(scratch, static_cast<int32_t>(char_code));
       __ St_h(scratch, FieldMemOperand(result_string,
                                        OFFSET_OF_DATA_START(SeqTwoByteString)));
     }
